Move utility/*.c to the root
