+++ Makefile
@@ -1,13 +1,5 @@
-# makefile for BOHM
+YFLAGS = -dv
 
-CFLAGS = -O 
-#CFLAGS = -g
-LDLIBS = -ll -ly
-YFLAGS = -d
-
-# No user serviceable parts below this line
-
-# each time one of the following is modified, everything is recompiled
 SUPPORT = \
 	e/crashhandler.e \
 	e/dynallhandler.e \
@@ -55,108 +47,58 @@ SRC = \
 	sthandler.c \
 	menu.c \
 	inspect.c \
-	utility/crashhandler.c \
-	utility/errorhandler.c \
-	utility/numberhandler.c \
-	utility/dynallhandler.c \
-	utility/filehandler.c \
-	utility/stringhandler.c
+	crashhandler.c \
+	errorhandler.c \
+	numberhandler.c \
+	dynallhandler.c \
+	filehandler.c \
+	stringhandler.c
 
 YLSRC = $(SRC) \
 	lambda_parser.y \
 	lambda_lexan.l
 
 GENSRC = $(SRC) \
-	y.tab.c \
-	lex.yy.c
+	lambda_parser.c \
+	lambda_lexan.c
 
 OBJS = \
-	main.o \
-	y.tab.o \
-	lex.yy.o \
-	copy.o \
-	sthandler.o \
-	scope_analysis.o \
-	inspect.o \
-	graphgenerator.o \
-	destroyer.o \
-	loader.o \
-	reducer.o \
-	m_stack.o \
-	readback.o \
-	garbage.o \
-	menu.o \
-	save.o \
-	utility/crashhandler.o \
-	utility/errorhandler.o \
-	utility/numberhandler.o \
-	utility/dynallhandler.o \
-	utility/filehandler.o \
-	utility/stringhandler.o
+	lib.a(lambda_parser.o) \
+	lib.a(lambda_lexan.o) \
+	lib.a(main.o) \
+	lib.a(copy.o) \
+	lib.a(sthandler.o) \
+	lib.a(scope_analysis.o) \
+	lib.a(inspect.o) \
+	lib.a(graphgenerator.o) \
+	lib.a(destroyer.o) \
+	lib.a(loader.o) \
+	lib.a(reducer.o) \
+	lib.a(m_stack.o) \
+	lib.a(readback.o) \
+	lib.a(garbage.o) \
+	lib.a(menu.o) \
+	lib.a(save.o) \
+	lib.a(crashhandler.o) \
+	lib.a(errorhandler.o) \
+	lib.a(numberhandler.o) \
+	lib.a(dynallhandler.o) \
+	lib.a(filehandler.o) \
+	lib.a(stringhandler.o)
 
-opt: $(OBJS) $(SUPPORT)
-	$(CC) $(CFLAGS) $(CPPFLAGS) $(LDFLAGS) $(OBJS) \
-	-o opt $(LDLIBS)
+all: opt
 
-# The two following productions are identical
-y.tab.c: lambda_parser.y
-	$(YACC) $(YFLAGS) lambda_parser.y
+debug:
+	$(MAKE) CFLAGS="-g -Wall -Werror"
 
-y.tab.h: lambda_parser.y
-	$(YACC) $(YFLAGS) lambda_parser.y
+opt: lib.a
+	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ lib.a -ll -ly
 
-lex.yy.c: lambda_lexan.l y.tab.h
-	$(LEX) $(LFLAGS) lambda_lexan.l
+lib.a: $(OBJS)
 
-# the default .c.o rule puts its object in the wrong directory
-.c.o:
-	$(COMPILE.c) -o $@ $<
+$(OBJS): $(SUPPORT)
 
-copy.o: copy.c $(SUPPORT)
-
-destroyer.o: destroyer.c $(SUPPORT)
-
-garbage.o: garbage.c $(SUPPORT)
-
-graphgenerator.o: graphgenerator.c $(SUPPORT)
-
-inspect.o: inspect.c $(SUPPORT)
-
-lex.yy.o: lex.yy.c $(SUPPORT) y.tab.h
-
-loader.o: loader.c $(SUPPORT)
-
-m_stack.o: m_stack.c $(SUPPORT)
-
-main.o: main.c $(SUPPORT)
-
-menu.o: menu.c $(SUPPORT)
-
-#rback.o: rback.c $(SUPPORT)
-
-readback.o: readback.c $(SUPPORT)
-
-reducer.o: reducer.c $(SUPPORT)
-
-save.o: save.c $(SUPPORT)
-
-scope_analysis.o: scope_analysis.c $(SUPPORT)
-
-sthandler.o: sthandler.c $(SUPPORT) y.tab.h
-
-y.tab.o: y.tab.c $(SUPPORT)
-
-utility/crashhandler.o: utility/crashhandler.c $(SUPPORT)
-
-utility/dynallhandler.o: utility/dynallhandler.c $(SUPPORT)
-
-utility/errorhandler.o: utility/errorhandler.c $(SUPPORT)
-
-utility/filehandler.o: utility/filehandler.c $(SUPPORT)
-
-utility/numberhandler.o: utility/numberhandler.c $(SUPPORT)
-
-utility/stringhandler.o: utility/stringhandler.c $(SUPPORT)
+y.tab.h: lambda_parser.c
 
 # lint the whole code.  Expect a lot of warnings!
 lint: $(GENSRC)
@@ -171,6 +113,6 @@ ctags:
 	ctags $(YLSRC)
 
 clean:
-	-rm -f y.tab.h y.tab.c lex.yy.c
-	-rm -f opt *.o utility/*.o
+	-rm -f opt lib.a *.o
+	-rm -f lambda_parser.c lambda_lexan.c y.*
 	-rm -f TAGS tags
