+++ stringhandler.c
@@ -19,8 +19,8 @@
 /* 1. Inclusion of header files.				*/
 /****************************************************************/
 
-#include		"../h/const.h"
-#include		"../h/types.h"
+#include		"h/const.h"
+#include		"h/types.h"
 
 
 /****************************************************************/
